requirements.txt
requests
pandas
numpy
scikit-learn
tensorflow
torch
scapy
elasticsearch
kibana
grafana
docker
pytest
jupyter
matplotlib
seaborn
plotly
networkx
pyshark
psutil
watchdog
loguru
pyarrow
torch_geometric
streamlit
Flask-SocketIO
python-socketio
websocket-client
